Add class to bake object texture to an image
That allows getting the voxel colors from any type of texture,
procedural or not, without having to evaluate the material node trees.
The baked texture image can be mapped to the voxels in the same way as
is currently done for image texture nodes from materials. The baking
process is configured to use only the color information, without
accounting for the lighting.

Since only the Cycles render engine can do baking, the class changes
to it before baking, and restores the previous engine afterwards. It
also has to add an image texture node to each material used by the
object, and those nodes can be removed afterwards as well. The image
is not yet removed on cleanup since the writer is going to use it
directly, but eventually this can be refactored to collect the color
data into a numpy array so the image can be deleted immediately to
ensure it does not linger in the .blend file if something fails.