Bake texture to an image to get voxel colors
Instead of manually evaluating the node trees to look for images or
solid colors, this code uses the Cycles renderer to bake the resulting
texture after all materials are applied, and gets the voxel colors from
that. For now, minimal changes have been made to the writer script to
include that functionality, so a lot code that is no longer needed is
still there. To be evaluated if that code can be removed or if it
should be kept to handle the case of simple node trees without
procedural textures, since the baking process can take long in some
cases.

Note that the code is not yet very stable. Sometimes it finds NaN for
the UV coordinates, and thus fails to export the object.